assets/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.0.0
